requirements-prod.txt
# Core Dependencies
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=20.1.0
python-dotenv>=0.19.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# NLP & Analysis
nltk>=3.6.0
textblob>=0.17.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Production Utilities
Werkzeug>=2.3.0
Click>=8.0.0
Jinja2>=3.0.0
MarkupSafe>=2.0.0

# Optional: For advanced deployment
# uwsgi>=2.0.0
# whitenoise>=6.0.0  # For serving static files
